Working version of websockets

+++ src/main/kotlin/com/dllewellyn/coinbaseapi/models/CurrencyBuyAndSell.kt
@@ -1,3 +1,3 @@
 package com.dllewellyn.coinbaseapi.models
 
-data class CurrencyBuyAndSell(val currencyFrom : String, val currencyTo : String, val from : Double, val to : Double)
+data class CurrencyBuyAndSell(val currencyFrom : String, val currencyTo : String, val from : Float, val to : Float)
