+++ src/main/kotlin/com/dllewellyn/coinbaseapi/models/CurrencyValue.kt
@@ -1,3 +1,3 @@
 package com.dllewellyn.coinbaseapi.models
 
-data class CurrencyValue(val currencyFrom : String, val currencyTo : String, val amount : Double)
+data class CurrencyValue(val currencyFrom : String, val currencyTo : String, val amount : Float)
